Added a simple action page with cards and redirects to the user profile

+++ app/controllers/sessions_controller.rb
@@ -7,8 +7,12 @@ class SessionsController < ApplicationController
         user = User.find_by(email: params[:email])
         if user != nil && user.authenticate(params[:password])
             session[:current_user_id] = user.id
-            redirect_to user
-            flash[:notice] = 'Welcome back.'
+            if current_user.admin?
+                redirect_to users_path
+            else
+                redirect_to user
+                flash[:notice] = 'Welcome back.'
+            end
           else
             redirect_to log_in_path
             flash[:notice] = 'E-mail and/or password is incorrect.'
@@ -16,9 +20,11 @@ class SessionsController < ApplicationController
     end
 
     def destroy
-        session.delete(:current_user_id)
-        @_current_user = nil
-        redirect_to root_path
-        flash[:notice] = 'Goodbye.'
+        if helpers.current_user.admin? == false || helpers.current_user != nil
+            session.delete(:current_user_id)
+            @_current_user = nil
+            redirect_to root_path
+            flash[:notice] = 'Goodbye.'
+        end
     end
 end
